Move controllers into src/index
