+++ src/models/cooler.js
