+++ src/views/template.js
